Refactor exterior
