+++ geopolars/src/lib.rs
@@ -2,6 +2,7 @@ pub mod error;
 pub mod geoarrow;
 pub mod geodataframe;
 pub mod geoseries;
+pub mod ops;
 pub mod spatial_index;
 pub mod util;
